+++ trunk/devel/libevbase/src/evpoll.h
@@ -11,7 +11,7 @@ int evpoll_update(EVBASE *evbase, EVENT *event);
 /* Delete event from evbase */
 int evpoll_del(EVBASE *evbase, EVENT *event);
 /* Loop evbase */
-int evpoll_loop(EVBASE *evbase, short, struct timeval *tv);
+int evpoll_loop(EVBASE *evbase, int, struct timeval *tv);
 /* Reset evbase */
 void evpoll_reset(EVBASE *evbase);
 /* Clean evbase */
